requirements.txt
gridstatusio==0.8.0
pandas==2.0.0
pytz==2023.3
matplotlib==3.7.1
seaborn==0.12.2
pyyaml==6.0.1
python-dotenv==1.0.0
atproto==0.0.2
loguru==0.7.0